feat: correct data type, user_no must be a Long

+++ thismuch/src/main/java/com/project/thismuch/data/dto/UserDTO.java
@@ -18,7 +18,7 @@ import lombok.extern.slf4j.Slf4j;
 @ToString
 @Builder
 public class UserDTO {
-    private Integer userNo; // pk
+    private Long userNo; // pk
     private String id;
     private String name;
     private String passwd;
